Move mapToRemandRequest files into mappers directory
