+++ packages/core/lib/policeGateway/leds/mappers/mapToRemandRequest/index.ts
@@ -1,0 +1,1 @@
+export { default as mapToRemandRequest } from "./mapToRemandRequest"
